Move font files back to src folder inside client
