+++ client/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import GlobalFonts from '../.fonts/Fonts'
+import GlobalFonts from './.fonts/Fonts'
 import App from 'components/App/App'
 import GlobalStyle from 'components/GlobalStyle'
 import { BrowserRouter as Router } from 'react-router-dom'
